Fixed bug that prevented the window from rendering when WX is between 0 and 6.

+++ Include/Graphics/PPU.hpp
@@ -59,7 +59,7 @@ namespace SHG
 
 		std::vector<Sprite> spritesOnCurrentScanline;
 		std::queue<Pixel> queuedSpritePixels;
-		uint8_t currentScanlineX = 0;
+		int16_t currentScanlineX = 0;
 		Memory& memoryMap;
 		Framebuffer primaryFramebuffer;
 		Framebuffer tileDebugFramebuffer;
